#1838 fix for windows platform
Wxwidgets color picker dialog doesn't support transparency/alpha on
windows. The performane with alpha channels rendering of wxWidgets is
also not good on Windows. The fix is add menu option "Fill Transparency
for Category" for windows only. And user needs to enable "Transparency
rendering for Windows" from preference dialog to make this function
working.

+++ Explore/SimpleHistCanvas.cpp
@@ -266,7 +266,7 @@ x_axis(0), y_axis(0), display_stats(false), show_axes(show_axes_),
 obs_id_to_ival(X.size())
 {
 	using namespace Shapefile;
-	for (size_t i=0, sz=X.size(); i<sz; i++) {
+	for (int i=0, sz=X.size(); i<sz; i++) {
         if (X_undef[i])
             continue;
         data_sorted.push_back(std::make_pair(X[i], i));
